+++ config/index.js
@@ -1,4 +1,5 @@
 import auth from "./auth";
 import database from "./database";
+import mail from "./mail";
 
-export { auth, database };
+export { auth, database, mail };